config/requirements.txt
streamlit>=1.31.0
openai>=1.12.0
python-dotenv>=1.0.0
langchain>=0.1.9
langchain-openai>=0.0.5
langchain-community>=0.0.20
faiss-cpu>=1.7.4
pypdf2>=3.0.1
python-docx>=1.1.0
requests>=2.31.0
tiktoken>=0.5.2
pandas>=2.0.0
numpy>=1.24.0

# Advanced parsers
pymupdf>=1.23.0
boto3>=1.34.0
Pillow>=10.0.0
docling>=1.0.0

# OCR dependencies for Docling
easyocr>=1.7.0
onnxruntime>=1.16.0

# Vector store backends
opensearch-py>=2.0.0
requests-aws4auth>=1.1.0

# FastAPI for REST API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pytest>=7.4.0

